requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
mysql-connector-python==8.1.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
requests==2.31.0
markdown==3.5
Werkzeug==2.3.7
email-validator==2.0.0
wtforms==3.0.1
PyMySQL==1.1.0
python-dateutil==2.8.2 